创建枚举类型便于后续代码编写

+++ src/main/java/com/evan/evanoj/service/QuestionSubmitService.java
@@ -18,5 +18,5 @@ public interface QuestionSubmitService extends IService<QuestionSubmit> {
      * @param loginUser
      * @return
      */
-    int doQuestionSubmit(QuestionSubmitAddRequest questionSubmitAddRequest, User loginUser);
+    long doQuestionSubmit(QuestionSubmitAddRequest questionSubmitAddRequest, User loginUser);
 }
